cypress: add redline_helper.js
