+++ cypress_test/integration_tests/desktop/writer/track_changes_spec.js
@@ -2,6 +2,7 @@
 
 var helper = require('../../common/helper');
 const desktopHelper = require('../../common/desktop_helper');
+const redlineHelper = require('../../common/redline_helper');
 
 describe(['tagdesktop', 'tagnextcloud', 'tagproxy'], 'Track Changes', function () {
 
@@ -12,30 +13,6 @@ describe(['tagdesktop', 'tagnextcloud', 'tagproxy'], 'Track Changes', function (
 		cy.cGet('#toolbar-up [id^="sidebar"] button:visible').click();
 		desktopHelper.selectZoomLevel('50', false);
 	});
-
-	function confirmChange(action) {
-		cy.cGet('#menu-editmenu').click();
-		cy.cGet('#menu-changesmenu').should($el => { expect(Cypress.dom.isDetached($el)).to.eq(false); }).click();
-		cy.cGet('#menu-changesmenu').should($el => { expect(Cypress.dom.isDetached($el)).to.eq(false); }).contains(action).click();
-		cy.cGet('body').type('{esc}');
-		cy.cGet('#menu-changesmenu').should('not.be.visible');
-	}
-
-	//enable record for track changes
-	function enableRecord() {
-		cy.cGet('#menu-editmenu').click();
-		cy.cGet('#menu-changesmenu').click();
-		cy.cGet('#menu-changesmenu').contains('Record').click();
-
-		cy.cGet('body').type('{esc}');
-
-		cy.cGet('#menu-editmenu').click();
-		cy.cGet('#menu-changesmenu').click();
-		cy.cGet('#menu-changesmenu').contains('Record').should('have.class', 'lo-menu-item-checked');
-
-		//to close
-		cy.cGet('#menu-changesmenu').click();
-	}
 
 	it('Accept All', function () {
 		helper.typeIntoDocument('Hello World');
@@ -48,7 +25,7 @@ describe(['tagdesktop', 'tagnextcloud', 'tagproxy'], 'Track Changes', function (
 			// Wait for animation
 			cy.wait(500);
 		}
-		enableRecord();
+		redlineHelper.enableRecord();
 
 		desktopHelper.getCompactIconArrow('DefaultNumbering').click();
 		cy.cGet('#insertannotation').click();
@@ -65,7 +42,7 @@ describe(['tagdesktop', 'tagnextcloud', 'tagproxy'], 'Track Changes', function (
 		cy.cGet('#comment-container-2').should('contain','some text1');
 		cy.cGet('div.cool-annotation').should('have.length', 3);
 
-		confirmChange('Accept All');
+		redlineHelper.confirmChange('Accept All');
 		cy.cGet('#comment-container-1').should('contain','some text0');
 		cy.cGet('#comment-container-2').should('not.exist');
 		cy.cGet('#comment-container-3').should('contain','some text2');
@@ -74,7 +51,7 @@ describe(['tagdesktop', 'tagnextcloud', 'tagproxy'], 'Track Changes', function (
 		helper.clearAllText();
 		helper.selectAllText();
 		cy.wait(500);
-		confirmChange('Accept All');
+		redlineHelper.confirmChange('Accept All');
 		helper.typeIntoDocument('{ctrl}a');
 		cy.wait(500);
 		helper.textSelectionShouldNotExist();
@@ -92,7 +69,7 @@ describe(['tagdesktop', 'tagnextcloud', 'tagproxy'], 'Track Changes', function (
 			// Wait for animation
 			cy.wait(500);
 		}
-		enableRecord();
+		redlineHelper.enableRecord();
 
 		desktopHelper.getCompactIconArrow('DefaultNumbering').click();
 		cy.cGet('#insertannotation').click();
@@ -109,7 +86,7 @@ describe(['tagdesktop', 'tagnextcloud', 'tagproxy'], 'Track Changes', function (
 		cy.cGet('#comment-container-2').should('contain','some text1');
 		cy.cGet('div.cool-annotation').should('have.length', 3);
 
-		confirmChange('Reject All');
+		redlineHelper.confirmChange('Reject All');
 		cy.cGet('#comment-container-1').should('contain','some text0');
 		cy.cGet('#comment-container-2').should('contain','some text1');
 		cy.cGet('#comment-container-2').should('not.have.class','tracked-deleted-comment-show');
@@ -119,7 +96,7 @@ describe(['tagdesktop', 'tagnextcloud', 'tagproxy'], 'Track Changes', function (
 		helper.clearAllText();
 		helper.selectAllText();
 		cy.wait(500);
-		confirmChange('Reject All');
+		redlineHelper.confirmChange('Reject All');
 		cy.cGet('#document-container').click();
 		helper.selectAllText();
 		cy.wait(500);
@@ -395,7 +372,7 @@ describe(['tagdesktop', 'tagnextcloud', 'tagproxy'], 'Track Changes', function (
 			// Wait for animation
 			cy.wait(500);
 		}
-		enableRecord();
+		redlineHelper.enableRecord();
 
 		desktopHelper.getCompactIconArrow('DefaultNumbering').click();
 		cy.cGet('#insertannotation').click();
